táblában javítva width típusa

+++ database/migrations/2024_03_15_181724_create_keyboards_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string("name", 100)->unique();
             $table->string("type", 30);
             $table->string("layout", 30);
-            $table->integer("width");
+            $table->float("width");
             $table->enum("wireless", ["yes","no"]);
             $table->timestamps("");
         });
